Split fastpath IntegerArray constructor and general purpose constructor (#22070)

+++ pandas/tests/extension/base/getitem.py
@@ -213,7 +213,7 @@ class BaseGetitemTests(BaseExtensionTests):
         s = pd.Series(data)
         result = s.take([0, -1])
         expected = pd.Series(
-            data._from_sequence([data[0], data[len(data) - 1]]),
+            data._from_sequence([data[0], data[len(data) - 1]], dtype=s.dtype),
             index=[0, len(data) - 1])
         self.assert_series_equal(result, expected)
